revisi lab 1

+++ ForMultiples20.java
@@ -14,7 +14,7 @@ public class ForMultiples20 {
             }
         }
 
-        avr = sum / counter;
+        avr = (double) sum / counter;
 
         System.out.printf("There are %d number that are multiple of %d in range 1 to 50.\n", counter, multiple);
         System.out.printf("The sum off all multiples of %d in range 1 to 50 is %d. \n", multiple, sum);
