+++ WhileMultiples20.java
@@ -15,7 +15,7 @@ public class WhileMultiples20 {
             i++;
         }
 
-        avr = sum / counter;
+        avr = (double) sum / counter;
 
         System.out.printf("There are %d number that are multiple of %d in range 1 to 50.\n", counter, multiple);
         System.out.printf("The sum off all multiples of %d in range 1 to 50 is %d. \n", multiple, sum);
